Ensure the scale is parsed as integers
This will fix strange problems with calculating the current value of the
slider.

For example:

If we pass the scale of ["25", "200"]. When the current position of the
handle is calculated, and the begining portion of the scale is added to
the calulated value we will have a integer + string. This will cause the
beigining portion of the scale to be added as a string instead of a
integer.

If the handle was at position 30 and we were calculating the positon it
would be 30 + "25" which will give us "3025" instead of 55 like we would
expect