fix(constants): only compute constants when needed
Instead of computing constants and accessing window as soon as the
module loads, compute constants when we need them. Only once.

This helps testing when using nouislider. Otherwise as soon as we
import this module, it tries to access window. which might not be
available given the environment. Yet. (= jsdom)